+++ src/commons/utils.ts
@@ -90,24 +90,44 @@ export const plainTextToHTML = (str: string): string => {
 	return '';
 };
 
+function isValidUrl(url: string): boolean {
+	const urlToCheck = /^(https?:\/\/)/.exec(url) ? url : `http://${url}`;
+	try {
+		const newUrl = new URL(urlToCheck);
+		return ['http:', 'https:'].includes(newUrl.protocol);
+	} catch (err) {
+		return false;
+	}
+}
+
 export const replaceLinkToAnchor = (content: string): string => {
 	if (content === '') {
 		return '';
 	}
-	const linkRegexp = /(https?:\/\/[^\s]+|www\.[^\s]+)/g;
+	const linkRegexp = new RegExp(
+		'(https?:\\/\\/)?' +
+			'(((?=([a-z\\d-@:]|[^\\u0000-\\u007F]){1,100}\\.)(xn--)?([a-z\\d]|[^\\u0000-\\u007F])+(-([a-z\\d]|[^\\u0000-\\u007F])+)*\\.)+[a-z]{2,100}|' +
+			'((\\d{1,3}\\.){3}\\d{1,3}))' +
+			'(:\\d+)?(/([-a-z\\d%_.~+#\\/()]|[^\\u0000-\\u007F])+)*(\\?[;&:@a-z\\d%_.~\\+=#\\-()]*)' +
+			'?([#:\\/\\.\\-a-z\\d_)]+)?',
+		'gi'
+	);
 	return content.replace(linkRegexp, (url) => {
-		const wrap = document.createElement('div');
-		const anchor = document.createElement('a');
+		if (isValidUrl(url)) {
+			const wrap = document.createElement('div');
+			const anchor = document.createElement('a');
 
-		const newInnerHtml = url.replace(/&#64;/g, '@').replace(/&#61;/g, '=');
-		let href = url;
-		if (!url.startsWith('http') && !url.startsWith('https')) {
-			href = `http://${url}`;
+			const newInnerHtml = url.replace(/&#64;/g, '@').replace(/&#61;/g, '=');
+			let href = url;
+			if (!url.startsWith('http') && !url.startsWith('https')) {
+				href = `http://${url}`;
+			}
+			anchor.href = href;
+			anchor.target = '_blank';
+			anchor.innerHTML = newInnerHtml;
+			wrap.appendChild(anchor);
+			return wrap.innerHTML;
 		}
-		anchor.href = href;
-		anchor.target = '_blank';
-		anchor.innerHTML = newInnerHtml;
-		wrap.appendChild(anchor);
-		return wrap.innerHTML;
+		return url;
 	});
 };