+++ src/views/search/list/conversation/search-conversation-list-item.tsx
@@ -17,13 +17,14 @@ import {
 	Text,
 	Tooltip
 } from '@zextras/carbonio-design-system';
-import { Tag, pushHistory, t, useUserAccounts, useUserSettings } from '@zextras/carbonio-shell-ui';
+import { pushHistory, t, useUserAccounts, useUserSettings } from '@zextras/carbonio-shell-ui';
 import { filter, forEach, includes, isEmpty, reduce, trimStart, uniqBy } from 'lodash';
 import styled from 'styled-components';
 
 import { SearchConversationMessagesList } from './search-conversation-messages-list';
 import { ZIMBRA_STANDARD_COLORS } from '../../../../carbonio-ui-commons/constants';
 import { useTags } from '../../../../carbonio-ui-commons/store/zustand/tags';
+import { Tag } from '../../../../carbonio-ui-commons/types/tags';
 import { participantToString } from '../../../../commons/utils';
 import { API_REQUEST_STATUS } from '../../../../constants';
 import { useConvPreviewOnSeparatedWindowFn } from '../../../../hooks/actions/use-conv-preview-on-separated-window';